+++ projects/worldskills-ng-auth/src/lib/services/auth.service.ts
@@ -5,8 +5,7 @@ import { User } from '../models/user';
 import { UserList } from '../models/user-list';
 import { GetUsersParams } from '../models/get-user-params';
 import { HttpUtil } from '../util/http.util';
-import { share } from 'rxjs/operators';
-import { GenericUtil } from '..//util/generic.util';
+import { GenericUtil } from '../util/generic.util';
 import { LIBRARY_CONFIG } from '../auth-lib-config';
 
 
@@ -21,35 +20,27 @@ export class AuthService {
 
   constructor() {
     this.appCode = this.config.api?.appCode || [];
-    this.endpoint = this.config.api?.apiEndpoint + '/auth' || '';
+    this.endpoint = this.config.api?.apiEndpoint ? `${this.config.api.apiEndpoint}/auth` : '';
   }
 
-  public getLoggedInUser(showCollapsedChildRoles: boolean = false): Observable<User> {
-    let params = new HttpParams();
-    params = params.set('show_child_roles', 'false');
-    params = params.set('show_collapsed_child_roles', String(showCollapsedChildRoles));
+  public getLoggedInUser(showCollapsedChildRoles: boolean = true): Observable<User> {
+    let params = new HttpParams()
+      .set('show_child_roles', 'false')
+      .set('show_collapsed_child_roles', String(showCollapsedChildRoles));
+
     if (!GenericUtil.isNullOrUndefined(this.appCode)) {
-      this.appCode.forEach(code => {
-        if (params.has('app_code')) {
-          params = params.append('app_code', String(code));
-        } else {
-          params = params.set('app_code', String(code));
-        }
-      });
+      this.appCode.forEach(code => params = params.append('app_code', String(code)));
     }
-    
-    const url = `${this.endpoint}/users/loggedIn`;
-    return this.http.get<User>(url, {params}).pipe(share());
+
+    return this.http.get<User>(`${this.endpoint}/users/loggedIn`, { params });
   }
 
   public logout(): Observable<any> {
-    const authUrl = `${this.endpoint}/sessions/logout`;
-    return this.http.post(authUrl, {});
+    return this.http.post(`${this.endpoint}/sessions/logout`, {});
   }
 
   public ping(): Observable<any> {
-    const url = `${this.endpoint}/ping`;
-    return this.http.get(url, {});
+    return this.http.get(`${this.endpoint}/ping`);
   }
 
   /*
@@ -58,18 +49,15 @@ export class AuthService {
 
   public listUsers(filter: GetUsersParams): Observable<UserList> {
     const params = HttpUtil.objectToParams(filter);
-    const url = `${this.endpoint}/users`;
-    return this.http.get<UserList>(url, {params});
+    return this.http.get<UserList>(`${this.endpoint}/users`, { params });
   }
 
   public getUser(id: number): Observable<User> {
-    const url = `${this.endpoint}/users/${id}`;
-    return this.http.get<User>(url);
+    return this.http.get<User>(`${this.endpoint}/users/${id}`);
   }
 
   public getUserByPerson(id: number): Observable<User> {
-    const url = `${this.endpoint}/users/person/${id}`;
-    return this.http.get<User>(url);
+    return this.http.get<User>(`${this.endpoint}/users/person/${id}`);
   }
 
   /*
@@ -77,22 +65,19 @@ export class AuthService {
   */
 
   public addRole(userId: number, roleId: number): Observable<any> {
-    const url = `${this.endpoint}/users/${userId}/roles`;
-    const data = { role_id: roleId };
-    return this.http.post(url, data);
+    return this.http.post(`${this.endpoint}/users/${userId}/roles`, { role_id: roleId });
   }
+
   public addRoleWithEntity(userId: number, roleId: number, entityId: number): Observable<any> {
-    const url = `${this.endpoint}/users/${userId}/roles`;
-    const data = { role_id: roleId, ws_entity_id: entityId };
-    return this.http.post(url, data);
+    return this.http.post(`${this.endpoint}/users/${userId}/roles`, { role_id: roleId, ws_entity_id: entityId });
   }
 
   public deleteRole(userId: number, roleId: number): Observable<any> {
-    const url = `${this.endpoint}/users/${userId}/roles/${roleId}`;
-    return this.http.delete(url);
+    return this.http.delete(`${this.endpoint}/users/${userId}/roles/${roleId}`);
   }
+
   public deleteRoleWithEntity(userId: number, roleId: number, entityId: number): Observable<any> {
-    const url = `${this.endpoint}/users/${userId}/roles/${roleId}?ws_entity_id=${entityId}`;
-    return this.http.delete(url);
+    const params = new HttpParams().set('ws_entity_id', String(entityId));
+    return this.http.delete(`${this.endpoint}/users/${userId}/roles/${roleId}`, { params });
   }
 }