+++ projects/worldskills-ng-tester-prime/src/environments/environment.ts
@@ -1,0 +1,7 @@
+export const environment = {
+  production: false,
+  apiUrl: 'https://api.worldskills.show',
+  loginUrl: 'https://auth.worldskills.show/oauth/authorize',
+  clientId: '2b74b4d695ca',
+  defaultUrl: 'http://localhost:10500/home',
+};